Refactor EB and ED modules to better accomodate other coordinate systems.
