+++ src/eb/relic_eb_neg.c
@@ -53,7 +53,7 @@ void eb_neg_basic(eb_t r, const eb_t p) {
 
 	fb_add(r->y, p->x, p->y);
 
-	r->norm = 1;
+	r->coord = BASIC;
 }
 
 #endif
@@ -70,14 +70,14 @@ void eb_neg_projc(eb_t r, const eb_t p) {
 		return;
 	}
 
-	if (p->norm) {
+	if (p->coord == BASIC) {
 		if (r != p) {
 			fb_copy(r->x, p->x);
 			fb_copy(r->z, p->z);
 		}
 
 		fb_add(r->y, p->x, p->y);
-		r->norm = 1;
+		r->coord = BASIC;
 		return;
 	}
 
@@ -91,7 +91,7 @@ void eb_neg_projc(eb_t r, const eb_t p) {
 			fb_copy(r->x, p->x);
 		}
 
-		r->norm = 0;
+		r->coord = PROJC;
 	}
 	CATCH_ANY {
 		THROW(ERR_CAUGHT);
