+++ src/eb/relic_eb_norm.c
@@ -48,7 +48,7 @@
  * @param inv			- the flag to indicate if z is already inverted.
  */
 static void eb_norm_imp(eb_t r, const eb_t p, int inv) {
-	if (!p->norm) {
+	if (p->coord != BASIC) {
 		if (inv) {
 			fb_copy(r->z, p->z);
 		} else {
@@ -60,7 +60,7 @@ static void eb_norm_imp(eb_t r, const eb_t p, int inv) {
 		fb_set_dig(r->z, 1);
 	}
 
-	r->norm = 1;
+	r->coord = BASIC;
 }
 
 #endif /* EB_ADD == PROJC */
@@ -75,7 +75,7 @@ static void eb_norm_hlv(eb_t r, const eb_t p) {
 	fb_add(r->y, p->x, p->y);
 	fb_mul(r->y, r->y, p->x);
 	fb_copy(r->x, p->x);
-	r->norm = 1;
+	r->coord = BASIC;
 }
 
 /*============================================================================*/
@@ -88,13 +88,13 @@ void eb_norm(eb_t r, const eb_t p) {
 		return;
 	}
 
-	if (p->norm == 1) {
+	if (p->coord == BASIC) {
 		/* If the point is represented in affine coordinates, just copy it. */
 		eb_copy(r, p);
 		return;
 	}
 
-	if (p->norm == 2) {
+	if (p->coord == HALVE) {
 		eb_norm_hlv(r, p);
 		return;
 	}
