+++ src/ed/relic_ed_cmp.c
@@ -24,59 +24,71 @@
 /**
  * @file
  *
- * Implementation of the Frobenius map on binary elliptic curves.
+ * Implementation of the Edwards elliptic curve point comparison.
  *
- * @ingroup eb
+ * @version $Id$
+ * @ingroup ed
  */
 
-#include "string.h"
+#include <assert.h>
 
 #include "relic_core.h"
-#include "relic_eb.h"
+#include "relic_label.h"
+#include "relic_md.h"
+
 
 /*============================================================================*/
 /* Public definitions                                                         */
 /*============================================================================*/
 
-#if defined(EB_KBLTZ)
+int ed_cmp(const ed_t p, const ed_t q) {
+    ed_t r, s;
+    int result = RLC_EQ;
 
-#if EB_ADD == BASIC || !defined(STRIP)
+    ed_null(r);
+    ed_null(s);
 
-void eb_frb_basic(eb_t r, const eb_t p) {
-	if (eb_is_infty(p)) {
-		eb_set_infty(r);
-		return;
-	}
+    TRY {
+        ed_new(r);
+        ed_new(s);
 
-	fb_sqr(r->x, p->x);
-	fb_sqr(r->y, p->y);
-	fb_zero(r->z);
-	fb_set_bit(r->z, 0, 1);
+        if ((!p->coord) && (!q->coord)) {
+            /* If the two points are not normalized, it is faster to compare
+             * x1 * z2 == x2 * z1 and y1 * z2 == y2 * z1. */
+            fp_mul(r->x, p->x, q->z);
+            fp_mul(s->x, q->x, p->z);
+            fp_mul(r->y, p->y, q->z);
+            fp_mul(s->y, q->y, p->z);
+#if ED_ADD == EXTND
+			fp_mul(r->t, p->t, q->z);
+			fp_mul(s->t, q->t, p->z);
+			if (fp_cmp(r->t, s->t) != RLC_EQ) {
+	            result = RLC_NE;
+	        }
+#endif
+        } else {
+			ed_copy(r, p);
+            ed_copy(s, q);
+            if (!p->coord) {
+                ed_norm(r, p);
+            }
+            if (!q->coord) {
+                ed_norm(s, q);
+            }
+        }
 
-	r->coord = BASIC;
+        if (fp_cmp(r->x, s->x) != RLC_EQ) {
+            result = RLC_NE;
+        }
+        if (fp_cmp(r->y, s->y) != RLC_EQ) {
+            result = RLC_NE;
+        }
+    } CATCH_ANY {
+        THROW(ERR_CAUGHT);
+    } FINALLY {
+        ep_free(r);
+        ep_free(s);
+    }
+
+    return result;
 }
-
-#endif
-
-#if EB_ADD == PROJC || !defined(STRIP)
-
-void eb_frb_projc(eb_t r, const eb_t p) {
-	if (eb_is_infty(p)) {
-		eb_set_infty(r);
-		return;
-	}
-
-	fb_sqr(r->x, p->x);
-	fb_sqr(r->y, p->y);
-	if (p->coord != BASIC) {
-		fb_sqr(r->z, p->z);
-	} else {
-		fb_set_dig(r->z, 1);
-	}
-
-	r->coord = p->coord;
-}
-
-#endif
-
-#endif
