+++ src/ed/relic_ed_dbl.c
@@ -76,7 +76,7 @@ void ed_dbl_basic(ed_t r, const ed_t p) {
 		fp_copy(r->x, t0);
 		fp_copy(r->z, p->z);
 
-		r->norm = 1;
+		r->coord = BASIC;
 	}
 	CATCH_ANY {
 		THROW(ERR_CAUGHT);
@@ -147,7 +147,7 @@ void ed_dbl_projc(ed_t r, const ed_t p) {
 		/* x3 = F * J */
 		fp_mul(r->z, t4, t6);
 
-		r->norm = 0;
+		r->coord = PROJC;
 	} CATCH_ANY {
 		THROW(ERR_CAUGHT);
 	} FINALLY {
@@ -210,14 +210,14 @@ void ed_dbl_extnd(ed_t r, const ed_t p) {
 		fp_mul(r->x, t2, t3);
 		/* Y = G * H */
 		fp_mul(r->y, t4, r->z);
-		if (r->norm != 2) {
+		if (r->coord != 2) {
 			/* T = E * H */
 			fp_mul(r->t, t2, r->z);
 		}
 		/* Z = F * G */
 		fp_mul(r->z, t3, t4);
 
-		r->norm = 0;
+		r->coord = PROJC;
 	} CATCH_ANY {
 		THROW(ERR_CAUGHT);
 	} FINALLY {
